move extended date value
