+++ stats/stats/src/charts/lines/new_blocks.rs
@@ -75,10 +75,10 @@ impl crate::Chart for NewBlocks {
 mod tests {
     use super::*;
     use crate::{
-        charts::db_interaction::{read::ExtendedDateValue, updater::get_min_block_blockscout},
+        charts::db_interaction::updater::get_min_block_blockscout,
         get_chart_data,
         tests::{init_db::init_db_all, mock_blockscout::fill_mock_blockscout_data},
-        Chart,
+        Chart, ExtendedDateValue,
     };
     use chrono::NaiveDate;
     use entity::chart_data;
